cleanup controller
