+++ EmployeeApp.Api/Services/IRefreshTokenGenerator.cs
@@ -1,7 +1,12 @@
-﻿namespace EmployeeApp.Api.Services
+﻿using EmployeeApp.Dal.Dtos;
+using System.Security.Claims;
+using System.Threading.Tasks;
+
+namespace EmployeeApp.Api.Services
 {
     public interface IRefreshTokenGenerator
     {
-        string GenerateToken();
+        Task<RefreshTokenDto> CreateTokenAndRefresh(string username, Claim[] claims);
+        Task AddRefreshToken(string username, string refreshToken);
     }
 }
